Re-enable mypy `warn_unused_ignores`
This took a long, long time to troubleshoot.
Ultimately, the issue is that `matplotlib` comes
with its own type stubs. It has to be installed for
the `type: ignore[attr-defined]` to be valid. It
must be installed in the `pre-commit` hook's
environment if we want it to pass the `nbqa` `mypy`
tests.